# src/docs/swagger.yaml
basePath: /
definitions:
...
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
...
  model.ReferralCode:
    properties:
      active:
        type: boolean
      code:
        type: string
      createdAt:
        type: string
      removedAt:
        type: string
      uid:
        type: string
    type: object
...
  model.Report:
    properties:
      active:
        type: boolean
      id:
        type: integer
      notes:
        type: string
      optId:
        type: integer
      reportedAt:
        type: string
      reportedUid:
        type: string
      resolvedAt:
        type: string
      uid:
        type: string
    type: object
  model.ReportExpanded:
    properties:
      active:
        type: boolean
      notes:
        type: string
      opt:
        type: string
      reportId:
        type: integer
      reportedAt:
        type: string
      reportedUid:
        type: string
      resolvedAt:
        type: string
      uid:
        type: string
    type: object
  model.ReportOpt:
    properties:
      active:
        type: boolean
      createdAt:
        type: string
      id:
        type: integer
      opt:
        type: string
      removedAt:
        type: string
    type: object
...
host: apidev.xopacks.com
info:
  contact: {}
  description: The official XOPAcks API
  title: XOPacks API
  version: "1.0"
paths:
...
  /referral/create:
    post:
      consumes:
      - application/json
      description: A creator can upload a new custom referral code to use for referring
        new creators
      parameters:
      - description: authorized uid
        in: query
        name: authorizedUid
        required: true
        type: string
      - description: referral code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.ReferralCode'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Create a custom code referral code
      tags:
      - Referral
  /referral/generate:
    post:
      consumes:
      - application/json
      description: A creator can generate a new referral code to use for referring
        new creators
      parameters:
      - description: authorized uid
        in: query
        name: authorizedUid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.ReferralCode'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Generate a new referral code
      tags:
      - Referral
  /referral/getActiveCodes:
    get:
      consumes:
      - application/json
      description: A creator can get a list of all active referral codes associated
        with them
      parameters:
      - description: authorized uid
        in: query
        name: authorizedUid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ReferralCode'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Get all active codes that a creator has
      tags:
      - Referral
  /referral/remove:
    delete:
      consumes:
      - application/json
      description: A creator can deactivate a referral code so it can no longer be
        used
      parameters:
      - description: authorized uid
        in: query
        name: authorizedUid
        required: true
        type: string
      - description: referral code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Remove a referral code
      tags:
      - Referral
  /report/opts:
    get:
      consumes:
      - application/json
      description: A user can get a list of the available report options
      parameters:
      - description: authorized uid
        in: query
        name: authorizedUid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ReportOpt'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Get active report options
      tags:
      - Report
  /report/submit:
    post:
      consumes:
      - application/json
      description: A user can submit a report against another user
      parameters:
      - description: authorized uid
        in: query
        name: authorizedUid
        required: true
        type: string
      - description: report object
        in: body
        name: report
        required: true
        schema:
          $ref: '#/definitions/model.Report'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.ReportExpanded'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Submit a new report
      tags:
      - Report
...
swagger: "2.0"